fix: respect `open` and `bonjour` for `node` API
