+++ lib/Server.js
@@ -11,6 +11,7 @@
 */
 const fs = require('fs');
 const path = require('path');
+const net = require('net');
 
 const ip = require('ip');
 const tls = require('tls');
@@ -35,8 +36,12 @@ const historyApiFallback = require('connect-history-api-fallback');
 const webpack = require('webpack');
 const webpackDevMiddleware = require('webpack-dev-middleware');
 
-const createLogger = require('./utils/createLogger');
+const createBonjour = require('./utils/createBonjour');
 const createCertificate = require('./utils/createCertificate');
+const createDomain = require('../lib/utils/createDomain');
+const createLogger = require('./utils/createLogger');
+const logStatus = require('./utils/logStatus');
+const open = require('./utils/open');
 
 const validateOptions = require('schema-utils');
 const schema = require('./options.json');
@@ -608,6 +613,73 @@ function Server (compiler, options = {}, _log) {
   websocketProxies.forEach(function (wsProxy) {
     this.listeningApp.on('upgrade', wsProxy.upgrade);
   }, this);
+
+  const suffix = (options.inline !== false || options.lazy === true ? '/' : '/webpack-dev-server/');
+
+  if (options.socket) {
+    this.listeningApp.on('error', (e) => {
+      if (e.code === 'EADDRINUSE') {
+        const clientSocket = new net.Socket();
+
+        clientSocket.on('error', (err) => {
+          if (err.code === 'ECONNREFUSED') {
+            // No other server listening on this socket so it can be safely removed
+            fs.unlinkSync(options.socket);
+
+            this.listen(options.socket, options.host, (error) => {
+              if (error) {
+                throw error;
+              }
+            });
+          }
+        });
+
+        clientSocket.connect({ path: options.socket }, () => {
+          throw new Error('This socket is already used');
+        });
+      }
+    });
+
+    this.listen(options.socket, options.host, (err) => {
+      if (err) {
+        throw err;
+      }
+      // chmod 666 (rw rw rw)
+      const READ_WRITE = 438;
+
+      fs.chmod(options.socket, READ_WRITE, (err) => {
+        if (err) {
+          throw err;
+        }
+
+        const uri = createDomain(options, this.listeningApp) + suffix;
+
+        logStatus(uri, options, this.log, options.stats.colors);
+
+        if (options.open) {
+          open(uri, options, this.log);
+        }
+      });
+    });
+  } else {
+    this.listen(options.port, options.host, (err) => {
+      if (err) {
+        throw err;
+      }
+
+      if (options.bonjour) {
+        createBonjour(options);
+      }
+
+      const uri = createDomain(options, this.listeningApp) + suffix;
+
+      logStatus(uri, options, this.log, options.stats.colors);
+
+      if (options.open) {
+        open(uri, options, this.log);
+      }
+    });
+  }
 }
 
 Server.prototype.use = function () {
